requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
python-dotenv
google-genai
gunicorn
supabase
pandas